requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
pycodestyle==2.6.0
python-dotenv==0.15.0
redis==3.5.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
